requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.1.7
astropy-iers-data==0.2025.1.20.0.32.27
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
av==12.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boxx==0.10.14
bpy==3.6.0
bpycv==0.4.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.2
colorlog==6.8.2
comm==0.2.2
ConfigArgParse==1.7
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.11
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
distlib==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
embreex==2.17.7.post5
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.0
gtsam==4.2
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
healpy==1.18.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.30.2
hydra-core==1.3.2
identify==2.6.0
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.2.33
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lit==18.1.8
lxml==5.3.0
manifold3d==2.5.1
mapbox_earcut==1.0.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathutils==3.3.0
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
minexr==1.0.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nerfacc @ git+https://github.com/KAIR-BAIR/nerfacc.git@433130618da036d64581e07dc1bf5520bd213129
# Editable install with no version control (nerfstudio==0.3.2)
# -e nvf_cvpr24
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
notebook_shim==0.2.4
numpy==1.26.4
nuscenes-devkit==1.1.11
nvidia-ml-py==12.535.161
nvitop==1.3.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.6.0.66
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plop==0.4.0
plotly==5.23.0
pluggy==1.5.0
portalocker==2.10.1
pre-commit==3.3.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pycocotools==2.0.8
pycollada==0.8
pycparser==2.22
pyequilib==0.5.8
pyerfa==2.0.1.5
pygame==2.6.1
Pygments==2.18.0
pyliblzfse==0.4.1
pymeshlab==2023.12.post1
pyngrok==7.2.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.9
pyparsing==3.1.4
pyquaternion==0.9.9
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-json-logger==2.0.7
python-socketio==5.11.3
pytorch3d==0.7.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
Rtree==1.3.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.13.0
setproctitle==1.3.3
Shapely==1.8.5.post1
shtab==1.7.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
snakeviz==2.2.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
svg.path==6.3
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.24
timm==1.0.15
tinycss2==1.3.0
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/@c91138bcd4c6877c8d5e60e483c0581aafc70cce#subdirectory=bindings/torch
tomli==2.0.1
torch==2.0.1+cu117
torch-fidelity==0.3.0
torchmetrics==1.4.1
torchvision==0.15.2+cu117
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trimesh==3.23.5
triton==2.0.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tyro==0.8.8
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
vhacdx==0.0.8.post1
virtualenv==20.26.3
viser @ git+https://github.com/nerfstudio-project/viser.git@aa417815bf248ba15ee6e22cd4bb49bbc149dee8
wandb==0.17.7
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wsproto==1.2.0
xatlas==0.0.9
xxhash==3.5.0
yacs==0.1.8
yourdfpy==0.0.56
zipp==3.20.0
zstandard==0.23.0
